+++ testflo/discover.py
@@ -10,7 +10,6 @@ from os.path import basename, dirname, isdir
 from testflo.util import find_files, get_module, ismethod
 from testflo.runner import get_testcase
 from testflo.result import TestResult
-from testflo.testspec import TestSpec
 
 
 class TestDiscoverer(object):
@@ -56,7 +55,7 @@ class TestDiscoverer(object):
         try:
             fname, mod = get_module(filename)
         except:
-            yield TestResult(TestSpec(filename), 0, 0, 'FAIL',
+            yield TestResult(filename, 0, 0, 'FAIL',
                              {'err_msg': traceback.format_exc()})
         else:
             if basename(fname).startswith(six.text_type('__init__.')):
@@ -82,7 +81,7 @@ class TestDiscoverer(object):
                 methods.append(''.join((fname, ':', testcase.__name__,
                                                '.', method.__name__)))
         for m in sorted(methods):
-            yield TestSpec(m, testcase)
+            yield m
 
     def _testspec_iter(self, testspec):
         """Iterate over expanded testspec strings found in the
