+++ testflo/result.py
@@ -3,15 +3,6 @@ import sys
 import time
 
 
-def elapsed_str(elapsed):
-    """return a string of the form hh:mm:sec"""
-    hrs = int(elapsed/3600)
-    elapsed -= (hrs * 3600)
-    mins = int(elapsed/60)
-    elapsed -= (mins * 60)
-    return "%02d:%02d:%.2f" % (hrs, mins, elapsed)
-
-
 class TestResult(object):
     """Contains the path to the test function/method, status
     of the test (if finished), error and stdout messages (if any),
@@ -34,7 +25,7 @@ class TestResult(object):
         """Returns the testspec with only the file's basename instead
         of its full path.
         """
-        parts = str(self.testspec).split(':', 1)
+        parts = self.testspec.split(':', 1)
         fname = os.path.basename(parts[0])
         return ':'.join((fname, parts[-1]))
 
@@ -45,140 +36,50 @@ class TestResult(object):
             return "%s: %s" % (self.testspec, self.status)
 
 
-class ResultPrinter(object):
-    """Prints the status and error message (if any) of each TestResult object
-    after its test has been run if verbose is True.  If verbose is False,
-    it displays a dot for each successful test, an 'S' for skipped tests,
-    and an 'F' for failed tests.  If a test fails, the error message is always
-    displayed, even in non-verbose mode.
+
+def _parse_test_path(testspec):
+    """Return a tuple of the form (fname, module, testcase, func)
+    based on the given testspec.
+
+    The format of testspec is one of the following:
+        <module>
+        <module>:<testcase>
+        <module>:<testcase>.<method>
+        <module>:<function>
+
+    where <module> is either the python module path or the actual
+    file system path to the .py file.  A value of None in the tuple
+    indicates that that part of the testspec was not present.
     """
 
-    def __init__(self, stream=sys.stdout, verbose=False):
-        self.stream = stream
-        self.verbose = verbose
-
-    def get_iter(self, input_iter):
-        for result in input_iter:
-            self._print_result(result)
-            yield result
-
-    def _print_result(self, result):
-        stream = self.stream
-
-        stats = elapsed_str(result.elapsed())
-        if result.memory_usage:
-            stats = stats + ', ' + str(result.memory_usage) + ' MB'
-
-        if self.verbose:
-            stream.write("%s ... %s (%s)\n%s" % (result.testspec,
-                                                 result.status,
-                                                 stats,
-                                                 result.err_msg))
-            if result.err_msg:
-                stream.write("\n")
-        elif result.status == 'OK':
-            stream.write('.')
-        elif result.status == 'FAIL':
-            stream.write('F')
-        elif result.status == 'SKIP':
-            stream.write('S')
-
-        if not self.verbose and result.err_msg:
-            if result.status == 'FAIL':
-                stream.write("\n%s ... %s (%s)\n%s\n" % (result.testspec,
-                                                         result.status,
-                                                         stats,
-                                                         result.err_msg))
-            elif result.status == 'SKIP':
-                stream.write("\n%s: SKIP: %s\n" % (result.short_name(),
-                                                   result.err_msg))
-
-        stream.flush()
-
-
-class ResultSummary(object):
-    """Writes a test summary after all tests are run."""
-
-    def __init__(self, options, stream=sys.stdout):
-        self.stream = stream
-        self.options = options
-        self._start_time = time.time()
-
-    def get_iter(self, input_iter):
-        oks = 0
-        total = 0
-        fails = []
-        skips = []
-        test_sum_time = 0.
-
-        write = self.stream.write
-
-        for test in input_iter:
-            total += 1
-
-            if test.status == 'OK':
-                oks += 1
-                test_sum_time += (test.end_time-test.start_time)
-            elif test.status == 'FAIL':
-                fails.append(test.short_name())
-                test_sum_time += (test.end_time-test.start_time)
-            elif test.status == 'SKIP':
-                skips.append(test.short_name())
-            yield test
-
-        # now summarize the run
-        if skips:
-            write("\n\nThe following tests were skipped:\n")
-            for s in sorted(skips):
-                write(s)
-                write('\n')
-
-        if fails:
-            write("\n\nThe following tests failed:\n")
-            for f in sorted(fails):
-                write(f)
-                write('\n')
+    testcase = method = None
+    testspec = str(testspec)
+    testspec = testspec.strip()
+    parts = testspec.split(':')
+    if len(parts) > 1 and parts[1].startswith('\\'):  # windows abs path
+        module = ':'.join(parts[:2])
+        if len(parts) == 3:
+            rest = parts[2]
         else:
-            write("\n\nOK")
+            rest = ''
+    else:
+        module, _, rest = testspec.partition(':')
 
-        write("\n\nPassed:  %d\nFailed:  %d\nSkipped: %d\n" %
-                            (oks, len(fails), len(skips)))
+    fname, mod = get_module(module)
 
-        wallclock = time.time() - self._start_time
-
-        s = "" if total == 1 else "s"
-        if self.options.isolated:
-            procstr = " in isolated processes"
+    if rest:
+        objname, _, method = rest.partition('.')
+        obj = getattr(mod, objname)
+        if inspect.isclass(obj) and issubclass(obj, unittest.TestCase):
+            testcase = obj
+            if method:
+                meth = getattr(obj, method)
+                if not ismethod(meth):
+                    raise TypeError("'%s' is not a method." % rest)
+        elif isinstance(obj, FunctionType):
+            method = obj
         else:
-            procstr = " using %d processes" % self.options.num_procs
-        write("\n\nRan %d test%s%s\nSum of test times: %s\n"
-              "Wall clock time:   %s\nSpeedup: %f\n\n" %
-                      (total, s, procstr,
-                       elapsed_str(test_sum_time), elapsed_str(wallclock),
-                       test_sum_time/wallclock))
+            raise TypeError("'%s' is not a TestCase or a function." %
+                            objname)
 
-
-class BenchmarkWriter(object):
-    """Writes benchmark data to a file for postprocessing.
-       Data is written as comma separated values (CSV)
-    """
-
-    def __init__(self, stream=sys.stdout):
-        self.timestamp = time.time()
-        self.stream = stream
-
-    def get_iter(self, input_iter):
-        for result in input_iter:
-            self._write_data(result)
-            yield result
-
-    def _write_data(self, result):
-        stream = self.stream
-        stream.write('%d,%s,%s,%f,%f\n' % (
-            self.timestamp,
-            result.testspec,
-            result.status,
-            result.elapsed(),
-            result.memory_usage
-        ))
-        stream.flush()
+    return (fname, mod, testcase, method)
